bot_telegram_otp/requirements.txt
# Bot Telegram OTP - Dependencias
python-telegram-bot==20.7
